% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpi_methods.R
\name{coef.scpi}
\alias{coef.scpi}
\title{Coef Method for Synthetic Control Methods}
\usage{
\method{coef}{scpi}(object, ...)
}
\arguments{
\item{object}{Class "scpi" object, obtained by calling \code{\link{scpi}}.}

\item{...}{Other arguments (eg. \code{ncols}).}
}
\value{
No return value, called to show \code{\link{scpi}} constructed weights.
}
\description{
The coef method for synthetic control prediction fitted objects.
}
\seealso{
\code{\link{scpi}} for synthetic control prediction.

Supported methods: \code{\link{print.scpi}}, \code{\link{summary.scpi}}, \code{\link{coef.scpi}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
