% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdataMulti_methods.R
\name{print.scdataMulti}
\alias{print.scdataMulti}
\title{Summary Method for Synthetic Control}
\usage{
\method{print}{scdataMulti}(x, ...)
}
\arguments{
\item{x}{Class "scdataMulti" object, obtained by calling  \code{\link{scdataMulti}}.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{scdataMulti}} results.
}
\description{
The print method for synthetic control data objects.
}
\seealso{
\code{\link{scdataMulti}} for synthetic control data preparation.

Supported methods: \code{\link{print.scdataMulti}}, \code{\link{summary.scdataMulti}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
