% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest_methods.R
\name{print.scest}
\alias{print.scest}
\title{Print Method for Synthetic Control Methods}
\usage{
\method{print}{scest}(x, ...)
}
\arguments{
\item{x}{Class "scest" object, obtained by calling \code{\link{scest}}.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{scest}} results.
}
\description{
The print method for synthetic control prediction fitted objects.
}
\seealso{
\code{\link{scest}} for synthetic control prediction.

Supported methods: \code{\link{print.scest}}, \code{\link{summary.scest}}, \code{\link{coef.scest}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
