% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpi_methods.R
\name{print.scpi}
\alias{print.scpi}
\title{Print Method for Synthetic Control Inference}
\usage{
\method{print}{scpi}(x, ...)
}
\arguments{
\item{x}{Class "scpi" object, obtained by calling \code{\link{scpi}}.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{scpi}} results.
}
\description{
The print method for synthetic control inference objects.
}
\seealso{
\code{\link{scpi}} for synthetic control inference

Supported methods: \code{\link{print.scpi}}, \code{\link{summary.scpi}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
