% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest.R
\name{scest}
\alias{scest}
\title{Estimation of Synthetic Control}
\usage{
scest(
  data,
  w.constr = NULL,
  V = NULL,
  opt.list = NULL,
  plot = FALSE,
  plot.name = NULL,
  plot.path = NULL,
  save.data = NULL
)
}
\arguments{
\item{data}{a class `scpi_data' object, obtained by calling \code{\link{scdata}}.}

\item{w.constr}{a list specifying the constraint set the estimated weights of the donors must belong to.
\code{w.constr} can contain up to four objects:
\itemize{
\item `\code{p}', a string indicating the norm to be constrained (\code{p} should be one of "no norm", "L1", and "L2")
\item `\code{dir}', a string indicating whether the constraint on the norm is an equality ("==") or inequality ("<=")
\item `\code{Q}', a scalar defining the value of the constraint on the norm
\item `\code{lb}', a scalar defining the lower bound on the weights. It can be either 0 or \code{-Inf}.
\item `\code{name}', a character selecting one of the default proposals.
See the \strong{Details} section for more.
}}

\item{V}{a weighting matrix to be used when minimizing the sum of squared
residuals
\deqn{(\mathbf{A}-\mathbf{B}\mathbf{w}-\mathbf{C}\mathbf{r})'\mathbf{V}(\mathbf{A}-\mathbf{B}\mathbf{w}-\mathbf{C}\mathbf{r})}
The default is the identity matrix, so equal weight is given to all observations.}

\item{opt.list}{a list specifying the stopping criteria and the algorithm for the underlying optimizer \code{\link{nloptr}} for point estimation.
See the \strong{Details} section for more.}

\item{plot}{a logical specifying whether \code{\link{scplot}} should be called and a plot saved in the current working directory. For more options see \code{\link{scplot}}.}

\item{plot.name}{a string containing the name of the plot (the format is by default .png). For more options see \code{\link{scplot}}.}

\item{plot.path}{a string containing the path at which the plot should be saved (default is output of \code{getwd()}.)}

\item{save.data}{a character specifying the name and the path of the saved dataframe containing the processed data used to produce the plot.}
}
\value{
\item{data}{object containing used data and some other values.}
\item{est.results}{object containing all the estimation results.}
}
\description{
The command constructs a counterfactual synthetic control unit as proposed in Cattaneo, Feng,
and Titiunik (2021).
}
\details{
\itemize{
\item{\strong{Estimation of Weights.} \code{w.constr} specifies the constraint set on the weights. First, the element
\code{p} allows the user to choose between imposing a constraint on either the L1 (\code{p = "L1"})
or the L2 (\code{p = "L2"}) norm of the weights and imposing no constraint on the norm (\code{p = "no norm"}).
Second, \code{Q} specifies the value of the constraint on the norm of the weights.
Third, \code{lb} sets the lower bound of each component of the vector of weights.
Fourth, \code{dir} sets the direction of the constraint on the norm in case \code{p = "L1"}
or \code{p = "L2"}. If \code{dir = "=="}, then
\deqn{||\mathbf{w}||_p = Q,\:\:\: w_j \geq lb,\:\: j =1,\ldots,J}
If instead \code{dir = "<="}, then
\deqn{||\mathbf{w}||_p \leq Q,\:\:\: w_j \geq lb,\:\: j =1,\ldots,J}
If instead \code{dir = "NULL"} no constraint on the norm of the weights is imposed.

An alternative to specifying an ad-hoc constraint set on the weights would be
choosing among some popular types of constraints. This can be done by including the element
`\code{name}' in the list \code{w.constr}. The following are available options:
\itemize{
\item {If \code{name == "simplex"} (the default), then
\deqn{||\mathbf{w}||_1 = 1,\:\:\: w_j \geq 0,\:\: j =1,\ldots,J.}}

\item {If \code{name == "lasso"}, then
\deqn{||\mathbf{w}||_1 \leq Q,}
where \code{Q} is by default equal to 1 but it can be provided as an element of the list (eg. \code{w.constr =
list(name = "lasso", Q = 2)}).}

\item{If \code{name == "ridge"}, then
\deqn{||\mathbf{w}||_2 \leq Q,}
where \code{Q} is a tuning parameter that is by default computed as
\deqn{(J+KM) \widehat{\sigma}_u^{2}/||\widehat{\mathbf{w}}_{OLS}||_{2}^{2}}
where \eqn{J} is the number of donors and \eqn{KM} is the total number of covariates used for adjustment.
The user can provide \code{Q} as an element of the list (eg. \code{w.constr =
list(name = "ridge", Q = 1)}).}

\item{If \code{name == "ols"}, then the problem is unconstrained and the vector of weights
is estimated via ordinary least squares.}
}
}

\item{\strong{Algorithm Options.} The default is a sequential quadratic programming (SQP) algorithm for nonlinearly constrained gradient-based optimization
(\code{algorithm = 'NLOPTR_LD_SLSQP'}) for all cases not involving the L1 norm.
For a complete list of algorithms see \href{https://nlopt.readthedocs.io/en/latest/NLopt_Algorithms/}{the official \code{nloptr} documentation}.
The other default values are \code{maxeval = 5000}, \code{ftol_res = 1.0e-8}, \code{ftol_abs = 1.0e-8}, \code{xtol_res = 1.0e-8}, \code{xtol_abs = 1.0e-8},
\code{tol_constraints_eq = 1.0e-8}, and \code{tol_constraints_ineq = 1.0e-8}. More information on the stopping criteria can be obtained running
\code{nloptr.print.options()} or \code{nloptr.get.default.options()}. f the optimization involves the L1 norm then \code{CVXR} is used for optimization.
More information on the stopping criteria can be obtained reading \href{https://cvxr.rbind.io/}{the official documentation}. }
}
}
\examples{

data <- scpi_germany

df <- scdata(df = data, id.var = "country", time.var = "year", 
             outcome.var = "gdp", period.pre = (1960:1990), 
             period.post = (1991:2013), unit.tr = "West Germany",
             unit.co = unique(data$country)[-7], constant = T,
             cointegrated.data = T)
             
result <- scest(df, w.constr = list(name = "simplex", Q = 1))
result <- scest(df, w.constr = list(lb = 0, dir = "==", p = "L1", Q = 1))


}
\references{
\itemize{
\item{\href{https://economics.mit.edu/files/17847}{Abadie, A. (2021)}. Using synthetic controls: Feasibility, data requirements, and methodological aspects.
\emph{Journal of Economic Literature}, 59(2), 391-425.}
\item{\href{https://cattaneo.princeton.edu/papers/Cattaneo-Feng-Titiunik_2021_JASA.pdf}{Cattaneo, M. D., Feng, Y., & Titiunik, R.
(2021)}. Prediction intervals for synthetic control methods. \emph{Journal of the American Statistical Association}, 116(536), 1865-1880.}
\item{\href{https://nppackages.github.io/references/Cattaneo-Feng-Palomba-Titiunik_2022_scpi.pdf}{Cattaneo, M. D., Feng, Y., Palomba F., and Titiunik, R. (2022).}.
scpi - Uncertainty Quantification for Synthetic Control Estimators.}}
}
\seealso{
\code{\link{scdata}}, \code{\link{scpi}}, \code{\link{scplot}}
}
\author{
\itemize{
\item{Matias Cattaneo, }{Princeton University}
\item{Yingjie Feng, }{Tsinghua University}
\item{Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.}
\item{Rocio Titiunik, Princeton University}}
}
