% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scpi_germany}
\alias{scpi_germany}
\title{Economic indicators for OECD countries}
\format{
A data frame with 748 rows and 11 variables:
\describe{
\item{index}{country index.}
\item{country}{name of the country.}
\item{year}{time index, in years.}
\item{gdp}{GDP per Capita (PPP, 2002 USD).}
\item{infrate}{annual percentage change in consumer prices (base year 1995).}
\item{trade}{trade openness measured as export plus imports as percentage of GDP.}
\item{schooling}{percentage of secondary school attained in the total population aged 25 and older.}
\item{industry}{industry share of value added.}
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/24714}
}
\usage{
scpi_germany
}
\description{
A dataset containing some economic indicators of 17 OECD countries from 1960
to 2003.
}
\keyword{datasets}
