% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdata_methods.R
\name{summary.scdata}
\alias{summary.scdata}
\title{Summary Method for Synthetic Control Prediction}
\usage{
\method{summary}{scdata}(object, ...)
}
\arguments{
\item{object}{Class "scest" object, obtained by calling \code{\link{scdata}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scdata}} results.
}
\description{
The summary method for synthetic control prediction objects.
}
\seealso{
\code{\link{scdata}}

Supported methods: \code{\link{print.scdata}}, \code{\link{summary.scdata}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
