% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdataMulti_methods.R
\name{summary.scdataMulti}
\alias{summary.scdataMulti}
\title{Summary Method for Synthetic Control Prediction}
\usage{
\method{summary}{scdataMulti}(object, ...)
}
\arguments{
\item{object}{Class "scdataMulti" object, obtained by calling \code{\link{scdataMulti}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scdataMulti}} results.
}
\description{
The summary method for synthetic control prediction objects.
}
\seealso{
\code{\link{scdataMulti}}

Supported methods: \code{\link{print.scdataMulti}}, \code{\link{summary.scdataMulti}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
