% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest_methods.R
\name{summary.scest}
\alias{summary.scest}
\title{Summary Method for Synthetic Control Prediction}
\usage{
\method{summary}{scest}(object, ...)
}
\arguments{
\item{object}{Class "scest" object, obtained by calling \code{\link{scest}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scest}} results.
}
\description{
The summary method for synthetic control prediction fitted objects.
}
\seealso{
\code{\link{scest}}

Supported methods: \code{\link{print.scest}}, \code{\link{summary.scest}}, \code{\link{coef.scest}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
