% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpi_methods.R
\name{summary.scpi}
\alias{summary.scpi}
\title{Summary Method for Synthetic Control Inference}
\usage{
\method{summary}{scpi}(object, ...)
}
\arguments{
\item{object}{Class "scpi" object, obtained by calling \code{\link{scpi}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scpi}} results.
}
\description{
The summary method for synthetic control inference objects.
}
\seealso{
\code{\link{scpi}}

Supported methods: \code{\link{print.scpi}}, \code{\link{summary.scpi}}.
}
\author{
Matias D. Cattaneo, Princeton University. \email{matias.d.cattaneo@gmail.com}.

Yingjie Feng, Tsinghua University. \email{fengyingjiepku@gmail.com}.

Filippo Palomba, Princeton University. \email{filippo.palomba19@gmail.com}.

Rocio Titiunik, Princeton University. \email{rocio.titiunik@gmail.com}.
}
