% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdata_methods.R
\name{print.scdata}
\alias{print.scdata}
\title{Summary Method for Synthetic Control}
\usage{
\method{print}{scdata}(x, ...)
}
\arguments{
\item{x}{Class "scdata" object, obtained by calling \code{\link{scdata}}.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{scdata}} results.
}
\description{
The print method for synthetic control data objects.
}
\seealso{
\code{\link{scdata}} for synthetic control data preparation.

Supported methods: \code{\link{print.scdata}}, \code{\link{summary.scdata}}.
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Yingjie Feng, Tsinghua University. \email{fengyj@sem.tsinghua.edu.cn}.

Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}.
}
