% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest_methods.R
\name{print.scest}
\alias{print.scest}
\title{Print Method for Synthetic Control Estimation}
\usage{
\method{print}{scest}(x, ...)
}
\arguments{
\item{x}{Class "scest" object, obtained by calling \code{\link{scest}}.}

\item{...}{Other arguments.}
}
\value{
No return value, called to print \code{\link{scest}} results.
}
\description{
The print method for synthetic control estimation objects.
}
\seealso{
\code{\link{scest}} for synthetic control estimation.

Supported methods: \code{\link{print.scest}}, \code{\link{summary.scest}}.
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Yingjie Feng, Tsinghua University. \email{fengyj@sem.tsinghua.edu.cn}.

Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}.
}
