% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpi_methods.R
\name{print.scpi}
\alias{print.scpi}
\title{Print Method for Synthetic Control Inference}
\usage{
\method{print}{scpi}(x, ...)
}
\arguments{
\item{x}{Class "scpi" object, obtained by calling \code{\link{scpi}}.}

\item{...}{Other arguments.}
}
\description{
The print method for synthetic control inference objects.
}
\seealso{
\code{\link{scpi}} for synthetic control inference

Supported methods: \code{\link{print.scpi}}, \code{\link{summary.scpi}}.
}
\author{
\itemize{
\item{Matias Cattaneo, }{Princeton University}
\item{Yingjie Feng, }{Tsinghua University}
\item{Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.}
\item{Rocio Titiunik, Princeton University}}
}
