% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpi-package.R
\docType{package}
\name{scpi-package}
\alias{scpi-package}
\alias{_PACKAGE}
\title{scpi: A package to compute Synthetic Control Prediction Intervals}
\description{
The package computes point estimates and prediction intervals for Synthetic Control methods.
}
\references{
\itemize{
\item{\href{https://cattaneo.princeton.edu/papers/Cattaneo-Feng-Titiunik_2021_JASA.pdf}{Cattaneo, M. D., Feng, Y., & Titiunik, R.
(2021)}. Prediction intervals for synthetic control methods. \emph{Journal of the American Statistical Association}, 116(536), 1865-1880.}
\item{\href{https://nppackages.github.io/references/Cattaneo-Feng-Palomba-Titiunik_2022_scpi.pdf}{Cattaneo, M. D., Feng, Y., Palomba F., and Titiunik, R. (2022)}}
}
}
\author{
\itemize{
\item{Matias Cattaneo, }{Princeton University}
\item{Yingjie Feng, }{Tsinghua University}
\item{Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.}
\item{Rocio Titiunik, Princeton University}}
}
