% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdata_methods.R
\name{summary.scdata}
\alias{summary.scdata}
\title{Summary Method for Synthetic Control Prediction}
\usage{
\method{summary}{scdata}(object, ...)
}
\arguments{
\item{object}{Class "scest" object, obtained by calling \code{\link{scdata}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scdata}} results.
}
\description{
The summary method for synthetic control prediction objects.
}
\seealso{
\code{\link{scdata}}

Supported methods: \code{\link{print.scdata}}, \code{\link{summary.scdata}}.
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Yingjie Feng, Tsinghua University. \email{fengyj@sem.tsinghua.edu.cn}.

Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}.
}
