% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest_methods.R
\name{summary.scest}
\alias{summary.scest}
\title{Summary Method for Synthetic Control Estimation}
\usage{
\method{summary}{scest}(object, ...)
}
\arguments{
\item{object}{Class "scest" object, obtained by calling \code{\link{scest}}.}

\item{...}{Additional arguments}
}
\description{
The summary method for synthetic control estimation objects.
}
\seealso{
\code{\link{scest}}

Supported methods: \code{\link{print.scest}}, \code{\link{summary.scest}}.
}
\author{
\itemize{
\item{Matias Cattaneo, }{Princeton University}
\item{Yingjie Feng, }{Tsinghua University}
\item{Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.}
\item{Rocio Titiunik, Princeton University}}
}
