% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scest_methods.R
\name{summary.scest}
\alias{summary.scest}
\title{Summary Method for Synthetic Control Estimation}
\usage{
\method{summary}{scest}(object, ...)
}
\arguments{
\item{object}{Class "scpi_scest" or "scpi_scest_multi" object, obtained by calling \code{\link{scest}}.}

\item{...}{Additional arguments}
}
\value{
No return value, called to summarize \code{\link{scest}} results.
}
\description{
The summary method for synthetic control estimation objects.
}
\seealso{
\code{\link{scest}}

Supported methods: \code{\link{print.scest}}, \code{\link{summary.scest}}.
}
\author{
Matias Cattaneo, Princeton University. \email{cattaneo@princeton.edu}.

Yingjie Feng, Tsinghua University. \email{fengyj@sem.tsinghua.edu.cn}.

Filippo Palomba, Princeton University (maintainer). \email{fpalomba@princeton.edu}.

Rocio Titiunik, Princeton University. \email{titiunik@princeton.edu}.
}
